+++ Source/SwiftLintFramework/Rules/LegacyConstantRule.swift
@@ -79,11 +79,13 @@ public struct LegacyConstantRule: CorrectableRule, ConfigurationProviderRule {
         let description = type(of: self).description
         var corrections = [Correction]()
         var contents = file.contents
-
         let matches = patterns.map({ pattern, template in
             file.matchPattern(pattern, withSyntaxKinds: [.identifier])
+                .filter { !file.ruleEnabledViolatingRanges([$0], forRule: self).isEmpty }
                 .map { ($0, pattern, template) }
         }).joined().sorted { $0.0.location > $1.0.location } // reversed
+
+        if matches.isEmpty { return [] }
 
         for (range, pattern, template) in matches {
             contents = regex(pattern).stringByReplacingMatches(in: contents,