+++ Source/SwiftLintFramework/Rules/LineLengthRule.swift
@@ -30,7 +30,7 @@ public struct LineLengthRule: ConfigurationProviderRule, SourceKitFreeRule {
     )
 
     public func validateFile(_ file: File) -> [StyleViolation] {
-        let minValue = configuration.params.map({$0.value}).min(by: <)
+        let minValue = configuration.params.map({ $0.value }).min(by: <)
         return file.lines.flatMap { line in
             // `line.content.characters.count` <= `line.range.length` is true.
             // So, `check line.range.length` is larger than minimum parameter value.