+++ Source/SwiftLintFramework/Rules/TrailingNewlineRule.swift
@@ -63,10 +63,10 @@ public struct TrailingNewlineRule: CorrectableRule, ConfigurationProviderRule, S
         guard let count = file.contents.trailingNewlineCount(), count != 1 else {
             return []
         }
-        let region = file.regions().filter {
-            $0.contains(Location(file: file.path, line: max(file.lines.count, 1)))
-        }.first
-        if region?.isRuleDisabled(self) == true {
+        guard let lastLineRange = file.lines.last?.range else {
+            return []
+        }
+        if file.ruleEnabledViolatingRanges([lastLineRange], forRule: self).isEmpty {
             return []
         }
         if count < 1 {