+++ Source/SwiftLintFramework/Rules/TrailingWhitespaceRule.swift
@@ -9,10 +9,10 @@
 import Foundation
 import SourceKittenFramework
 
-public struct TrailingWhitespaceRule: CorrectableRule, ConfigurationProviderRule,
-                                      SourceKitFreeRule {
+public struct TrailingWhitespaceRule: CorrectableRule, ConfigurationProviderRule {
 
-    public var configuration = TrailingWhitespaceConfiguration(ignoresEmptyLines: false)
+    public var configuration = TrailingWhitespaceConfiguration(ignoresEmptyLines: false,
+                                                               ignoresComments: true)
 
     public init() {}
 
@@ -20,17 +20,27 @@ public struct TrailingWhitespaceRule: CorrectableRule, ConfigurationProviderRule
         identifier: "trailing_whitespace",
         name: "Trailing Whitespace",
         description: "Lines should not have trailing whitespace.",
-        nonTriggeringExamples: [ "//\n" ],
-        triggeringExamples: [ "// \n" ],
-        corrections: [ "// \n": "//\n" ]
+        nonTriggeringExamples: [ "let name: String\n", "//\n", "// \n",
+            "let name: String //\n", "let name: String // \n" ],
+        triggeringExamples: [ "let name: String \n", "/* */ let name: String \n" ],
+        corrections: [ "let name: String \n": "let name: String\n",
+            "/* */ let name: String \n": "/* */ let name: String\n"]
     )
 
     public func validateFile(_ file: File) -> [StyleViolation] {
         let filteredLines = file.lines.filter {
-            $0.content.hasTrailingWhitespace() &&
-                (!configuration.ignoresEmptyLines ||
+            guard $0.content.hasTrailingWhitespace() else { return false }
+
+            let commentKinds = SyntaxKind.commentKinds()
+            if configuration.ignoresComments,
+                let lastSyntaxKind = file.syntaxKindsByLines[$0.index].last,
+                commentKinds.contains(lastSyntaxKind) {
+                return false
+            }
+
+            return !configuration.ignoresEmptyLines ||
                     // If configured, ignore lines that contain nothing but whitespace (empty lines)
-                    !$0.content.trimmingCharacters(in: .whitespaces).isEmpty)
+                    !$0.content.trimmingCharacters(in: .whitespaces).isEmpty
         }
 
         return filteredLines.map {
@@ -44,8 +54,20 @@ public struct TrailingWhitespaceRule: CorrectableRule, ConfigurationProviderRule
         let whitespaceCharacterSet = CharacterSet.whitespaces
         var correctedLines = [String]()
         var corrections = [Correction]()
-        let fileRegions = file.regions()
         for line in file.lines {
+            guard line.content.hasTrailingWhitespace() else {
+                correctedLines.append(line.content)
+                continue
+            }
+
+            let commentKinds = SyntaxKind.commentKinds()
+            if configuration.ignoresComments,
+                let lastSyntaxKind = file.syntaxKindsByLines[line.index].last,
+                commentKinds.contains(lastSyntaxKind) {
+                correctedLines.append(line.content)
+                continue
+            }
+
             let correctedLine = (line.content as NSString)
                 .trimmingTrailingCharacters(in: whitespaceCharacterSet)
 
@@ -54,10 +76,7 @@ public struct TrailingWhitespaceRule: CorrectableRule, ConfigurationProviderRule
                 continue
             }
 
-            let region = fileRegions.filter {
-                $0.contains(Location(file: file.path, line: line.index, character: 0))
-            }.first
-            if region?.isRuleDisabled(self) == true {
+            if file.ruleEnabledViolatingRanges([line.range], forRule: self).isEmpty {
                 correctedLines.append(line.content)
                 continue
             }